Making it more data efficient.

+++ Base/Source/Gathering of Components/ID_Component.h
@@ -7,7 +7,7 @@ class ID_Component : public GenericComponent
 {
 public:
     ID_Component() {}
-    ID_Component(const size_t &zeID) { zeID_ = zeID; }
+    ID_Component(const unsigned short &zeID) { zeID_ = zeID; }
     virtual ~ID_Component() {}
 
     virtual bool onNotify(const int &zeEvent) {
@@ -18,10 +18,10 @@ public:
         }
         return false;
     };
-    size_t &getValue() { return zeID_; }
+    unsigned short &getValue() { return zeID_; }
 
 protected:
-    size_t zeID_;
+    unsigned short zeID_;
 };
 
 #endif
